+++ docs/changelog.md
@@ -1,5 +1,18 @@
 # scrapy-playwright changelog
 
+### [v0.0.39](https://github.com/scrapy-plugins/scrapy-playwright/releases/tag/v0.0.39) (2024-07-11)
+
+* Return proper status and headers for downloads (#293)
+* Restart on browser crash (#295)
+* Override method and/or body only for the first matching request (#297)
+
+
+### [v0.0.38](https://github.com/scrapy-plugins/scrapy-playwright/releases/tag/v0.0.38) (2024-07-06)
+
+* Fix freezing on responses with status 204 (#292)
+* Connect to remote browser using BrowserType.connect (#283)
+
+
 ### [v0.0.37](https://github.com/scrapy-plugins/scrapy-playwright/releases/tag/v0.0.37) (2024-07-03)
 
 * Improve Windows concurrency (#286)